+++ tests/projects/dlang/dub_package/xmake.lua
@@ -1,8 +1,0 @@
-add_rules("mode.debug", "mode.release")
-
-add_requires("dub::log 0.4.3")
-
-target("test")
-    set_kind("binary")
-    add_files("src/main.d")
-    add_packages("dub::log")