+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * [#201](https://github.com/TOMO-CAT/xmake/issues/201): object target 默认编译隐式动态库, 提前暴露丢失的 syslinks 和 packages 信息
 * [#211](https://github.com/TOMO-CAT/xmake/issues/211): 支持 xmake_root.lua 确定 xmake 项目根目录
 * [#210](https://github.com/TOMO-CAT/xmake/issues/210): 新增 test kind, 简化单测描述域逻辑
+* [#207](https://github.com/TOMO-CAT/xmake/issues/207): xmake test 支持默认情况下运行 default test, 和 xmake run 保持统一
+* [#224](https://github.com/TOMO-CAT/xmake/issues/224): 在单测沙盒中建立单测二进制软链, 方便调试
 
 ### Bugs 修复
 
@@ -18,6 +20,7 @@
 * [#191](https://github.com/TOMO-CAT/xmake/issues/191): 修复单测耗时统计不准的问题
 * [#194](https://github.com/TOMO-CAT/xmake/issues/194): 修复 protobuf rule 可能丢失 proto 文件依赖顺序的问题
 * [#220](https://github.com/TOMO-CAT/xmake/issues/220): 修复依赖 shared target 时以全路径链接该动态库导致运行时难以加载动态库的问题
+* [#225](https://github.com/TOMO-CAT/xmake/issues/225): 包含动态库的 package 禁用 fullpath
 
 ## v3.0.6